job_portal/jobportal/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-allauth==0.60.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.14
python-magic==0.4.27
django-environ==0.11.2
dj-database-url==2.1.0
djangorestframework==3.14.0
scikit-learn==1.3.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.0.1

#note if error installation just remove the version